+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - :fire:GraphQL:fire: support with the new Graphene Bundle
 - add support for specifying parameters to inject into classes as class attributes
+- when using `unchained.inject()` on a class, or subclassing a class that supports automatic dependency injection, all non-dunderscore methods now support having dependencies injected
 
 #### Configuration Improvements